+++ engine/flink/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/source/flink/FlinkKafkaSource.scala
@@ -3,12 +3,12 @@ package pl.touk.nussknacker.engine.kafka.source.flink
 import cats.data.NonEmptyList
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.flink.api.common.functions.{OpenContext, RuntimeContext}
-import org.apache.flink.streaming.api.datastream.DataStreamSource
+import org.apache.flink.streaming.api.datastream.{DataStream, DataStreamSource, SingleOutputStreamOperator}
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
 import org.apache.flink.streaming.api.functions.source.SourceFunction
 import org.apache.flink.streaming.connectors.kafka.{FlinkKafkaConsumer, FlinkKafkaConsumerBase}
 import org.apache.kafka.clients.consumer.ConsumerRecord
-import pl.touk.nussknacker.engine.api.NodeId
+import pl.touk.nussknacker.engine.api.{Context, NodeId}
 import pl.touk.nussknacker.engine.api.component.StaticParameterConfig
 import pl.touk.nussknacker.engine.api.definition.{FixedExpressionValue, FixedValuesWithRadioParameterEditor, Parameter}
 import pl.touk.nussknacker.engine.api.deployment.{ScenarioActionName, WithActionParametersSupport}
@@ -22,18 +22,16 @@ import pl.touk.nussknacker.engine.api.process.{
 }
 import pl.touk.nussknacker.engine.api.runtimecontext.{ContextIdGenerator, EngineRuntimeContext}
 import pl.touk.nussknacker.engine.api.test.{TestData, TestRecord, TestRecordParser}
+import pl.touk.nussknacker.engine.flink.api.compat.ExplicitUidInOperatorsSupport
 import pl.touk.nussknacker.engine.flink.api.exception.ExceptionHandler
-import pl.touk.nussknacker.engine.flink.api.process.{
-  FlinkCustomNodeContext,
-  FlinkSourceTestSupport,
-  StandardFlinkSource,
-  StandardFlinkSourceFunctionUtils
-}
+import pl.touk.nussknacker.engine.flink.api.process._
 import pl.touk.nussknacker.engine.flink.api.timestampwatermark.{
   StandardTimestampWatermarkHandler,
   TimestampWatermarkHandler
 }
 import pl.touk.nussknacker.engine.flink.api.timestampwatermark.StandardTimestampWatermarkHandler.SimpleSerializableTimestampAssigner
+import pl.touk.nussknacker.engine.flink.watermarkstrategy.FlinkWatermarkStrategyRuntimeHandler
+import pl.touk.nussknacker.engine.flink.watermarkstrategy.FlinkWatermarkStrategyRuntimeHandler.ContextWithEventTime
 import pl.touk.nussknacker.engine.kafka._
 import pl.touk.nussknacker.engine.kafka.serialization.FlinkSerializationSchemaConversions
 import pl.touk.nussknacker.engine.kafka.serialization.FlinkSerializationSchemaConversions.FlinkDeserializationSchemaWrapper
@@ -45,6 +43,7 @@ import pl.touk.nussknacker.engine.kafka.source.flink.FlinkKafkaSource.{
 import pl.touk.nussknacker.engine.schemedkafka.KafkaUniversalComponentTransformer.inputParamName
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.UniversalToJsonFormatter
 import pl.touk.nussknacker.engine.util.parameters.TestingParametersSupport
+import pl.touk.nussknacker.engine.util.watermarkstrategy.WatermarkStrategyOptions
 
 import java.util
 import java.util.Properties
@@ -58,20 +57,69 @@ class FlinkKafkaSource[K, V](
     protected override val formatter: UniversalToJsonFormatter[K, V],
     override val contextInitializer: ContextInitializer[ConsumerRecord[K, V]],
     testParametersInfo: KafkaTestParametersInfo,
-    namingStrategy: NamingStrategy
-) extends StandardFlinkSource[ConsumerRecord[K, V]]
+    namingStrategy: NamingStrategy,
+    watermarkStrategyOptions: WatermarkStrategyOptions
+) extends FlinkSource
+    with ExplicitUidInOperatorsSupport
     with Serializable
+    // These mixins below are for scenario testing mechanism using source-specific test data format
     with FlinkSourceTestSupport[ConsumerRecord[K, V]]
     with TestDataGenerator
     with TestWithParametersSupport[ConsumerRecord[K, V]]
+    with CustomizableContextInitializerSource[ConsumerRecord[K, V]]
+    // end
     with WithActionParametersSupport
     with LazyLogging
     with KafkaLiveDataProvider[K, V] {
 
   private val typeInformation = ConsumerRecordTypeInfo[K, V](kafkaComponentsConfig)
 
+  override def contextStream(
+      env: StreamExecutionEnvironment,
+      flinkNodeContext: FlinkCustomNodeContext
+  ): DataStream[Context] = {
+    val streamOfRaw = sourceStream(env, flinkNodeContext)
+    // 1. set UID and override source name
+    val rawSourceWithUid = sourceWithUidAndName(streamOfRaw, flinkNodeContext)
+
+    // 2. initialize Context and compute event time
+    rawSourceWithUid
+      .map(
+        new FlinkWatermarkStrategyRuntimeHandler.ContextInitializingFunction(
+          contextInitializer,
+          flinkNodeContext.nodeId,
+          flinkNodeContext.convertToEngineRuntimeContext,
+          watermarkStrategyOptions.eventTimeLazyParam,
+          flinkNodeContext.lazyParameterHelper
+        ),
+        FlinkWatermarkStrategyRuntimeHandler.contextInitializingFunctionOutputTypeInfo(
+          flinkNodeContext.asOneOutputContext
+        )
+      )
+      // 3. assign timestamp and watermarks
+      .assignTimestampsAndWatermarks(
+        FlinkWatermarkStrategyRuntimeHandler.watermarkStrategy(watermarkStrategyOptions)
+      )
+      // 4. unwrap context
+      .map((ctxWithEventTime: ContextWithEventTime) => ctxWithEventTime.context, flinkNodeContext.contextTypeInfo)
+  }
+
+  private def sourceWithUidAndName[T](
+      streamOfRaw: DataStream[T],
+      flinkNodeContext: FlinkCustomNodeContext
+  ): DataStream[T] = {
+    streamOfRaw match {
+      case singleOut: SingleOutputStreamOperator[_] =>
+        setUidToNodeIdIfNeed[T](
+          flinkNodeContext,
+          singleOut.name(flinkNodeContext.nodeId)
+        )
+      case _ => streamOfRaw
+    }
+  }
+
   @nowarn("cat=deprecation")
-  override def sourceStream(
+  private def sourceStream(
       env: StreamExecutionEnvironment,
       flinkNodeContext: FlinkCustomNodeContext
   ): DataStreamSource[ConsumerRecord[K, V]] = {
@@ -187,15 +235,6 @@ class FlinkKafkaSource[K, V](
     Some(
       StandardTimestampWatermarkHandler.afterEachEvent[ConsumerRecord[K, V]](
         (_.timestamp()): SimpleSerializableTimestampAssigner[ConsumerRecord[K, V]]
-      )
-    )
-
-  override def timestampAssigner: Option[TimestampWatermarkHandler[ConsumerRecord[K, V]]] =
-    Some(
-      StandardTimestampWatermarkHandler.boundedOutOfOrderness(
-        extract = None,
-        maxOutOfOrderness = kafkaComponentsConfig.defaultMaxOutOfOrdernessMillis,
-        idlenessTimeoutDuration = kafkaComponentsConfig.idleTimeoutDuration
       )
     )